refactor: extract Pokémon listing logic into separate PokemonHome component for better organization and maintainability
